third_party/WebKit/LayoutTests/fast/canvas/webgl/texImage-imageBitmap-from-offscreen-canvas-resize.html

<!DOCTYPE html>
<html>
<head>
<script src="./resources/webgl-test-utils-full.js"></script>
<script src="./resources/tex-image-and-sub-image-utils.js"></script>
<script src="./resources/tex-image-and-sub-image-image-bitmap-utils-resize.js"></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
var wtu = WebGLTestUtils;
var tiu = TexImageUtils;
var gl = null;
var internalFormat = "RGBA";
var pixelFormat = "RGBA";
var pixelType = "UNSIGNED_BYTE";

function setCanvasToRedGreen(ctx) {
    var width = ctx.canvas.width;
    var halfWidth = Math.floor(width / 2);
    var height = ctx.canvas.height;
    var halfHeight = Math.floor(height / 2);
    ctx.fillStyle = "rgba(255, 0, 0, 1)";
    ctx.fillRect(0, 0, halfWidth, halfHeight);
    ctx.fillStyle = "rgba(255, 0, 0, 0.1)";
    ctx.fillRect(halfWidth, 0, halfWidth, halfHeight);
    ctx.fillStyle = "rgba(0, 255, 0, 1)";
    ctx.fillRect(0, halfHeight, halfWidth, halfHeight);
    ctx.fillStyle = "rgba(0, 255, 0, 0.1)";
    ctx.fillRect(halfWidth, halfHeight, halfWidth, halfHeight);
}

promise_test(function() {
    var bitmaps = [];

    var canvas = document.createElement('canvas');
    canvas.width = 4;
    canvas.height = 4;
    document.body.appendChild(canvas);
    gl = canvas.getContext("webgl");

    gl.clearColor(0,0,0,1);
    gl.clearDepth(1);

    var testCanvas = document.createElement('canvas');
    testCanvas.width = 2;
    testCanvas.height = 2;
    var offscreen = testCanvas.transferControlToOffscreen();
    var ctx = offscreen.getContext('2d');
    setCanvasToRedGreen(ctx);

    var options = {resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"};
    var p1 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.defaultOption = imageBitmap });

    options.imageOrientation = "none";
    options.premultiplyAlpha = "premultiply";
    var p2 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.noFlipYPremul = imageBitmap });

    options.premultiplyAlpha = "default";
    var p3 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.noFlipYDefault = imageBitmap });

    options.premultiplyAlpha = "none";
    var p4 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.noFlipYUnpremul = imageBitmap });

    options.imageOrientation = "flipY";
    options.premultiplyAlpha = "premultiply";
    var p5 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.flipYPremul = imageBitmap });

    options.premultiplyAlpha = "default";
    var p6 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.flipYDefault = imageBitmap });

    options.premultiplyAlpha = "none";
    var p7 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.flipYUnpremul = imageBitmap });

    options = {resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"};
    var p8 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.colorSpaceDef = imageBitmap });

    options.colorSpaceConversion = "none";
    var p9 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.colorSpaceNone = imageBitmap });

    options.colorSpaceConversion = "default";
    var p10 = createImageBitmap(offscreen, options).then(
        function(imageBitmap) { bitmaps.colorSpaceDefault = imageBitmap });

    return Promise.all([p1, p2, p3, p4, p5, p6, p7, p8, p9, p10]).then(function() {
        var alphaVal = 0.5;
        var testPassed = runTest(bitmaps, alphaVal, false);
        if (!testPassed)
            assert_true(false, 'Test failed');
    }, function() {
        assert_true(false, 'Promise rejected');
    });
}, 'createImageBitmap(HTMLCanvasElement) with resize and other options');
</script>
